Split seekers types into multiple files
