+++ requirements.txt
@@ -1,3 +1,4 @@
-pygame>=2.1.0
+pygame~=2.6.0
 grpcio==1.59.3
 protobuf==4.23.2
+contextvars==2.4
