+++ seekers/__init__.py
@@ -1,14 +1,4 @@
 from __future__ import annotations
 
 from .colors import Color
-from .seekers_types import (
-    Config,
-    Vector,
-    Physical,
-    Goal,
-    Magnet,
-    Seeker,
-    Player,
-    World,
-    Camp,
-)
+from .game import *
