+++ seekers/game/__init__.py
@@ -1,0 +1,9 @@
+from .vector import Vector
+from .config import Config
+from .camp import Camp
+from .seeker import Seeker, Magnet
+from .goal import Goal
+
+from .camp import Camp
+from .player import Player
+from .game import SeekersGame
