+++ seekers/game/game_logic.py
@@ -1,7 +1,13 @@
-from .draw import ScoreAnimation, Animation
-from .seekers_types import *
-
 import typing
+
+from .player import Player
+from .camp import Camp
+from .goal import Goal
+from .world import World
+from .vector import Vector
+from .seeker import Seeker
+from .physical import Physical
+from seekers.draw import ScoreAnimation, Animation
 
 
 def tick(players: typing.Iterable[Player], camps: list[Camp], goals: list[Goal],
